arcli: allow for more natural syntax on job deletion (abortion)
Currently to abort a job one would run:

$ arcli job -d -i <JOB_ID>

Where '-d' is the delete action, and the job to be deleted is job
whose ID is given with '-i'.

This coherent with the other commands that allow one to delete, say
a host, by either name or ID precisely. The problem is that is not
really straightforward, so allow a second syntax:

$ arcli job -d <JOB_ID>

Since a job will never be deleted/aborted based on its name, but
always based on its ID.

Signed-off-by: Cleber Rosa <crosa@redhat.com>