TKO support: introduce support for multiple services
This is not TKO-only feature, but an enabler to TKO and any other
service that may be added and supported in the feature.

Currently a connection object supports only a single service,
that is, 'AFE'. In fact the current concept of a connection has a
1:1 mapping to a service. This is bad because users would have to
keep various connection objects.

At the object API level, this would be even worse, with users having
to pass the right connection object to the right method.  So, to ease
things at the API level, RPC methods will know what service they
belong to, so that the users of the API do not have to keep track
of that information.

Signed-off-by: Cleber Rosa <crosa@redhat.com>