arc.shared: create a new top level module and namespace
That is intended to be used as a portable (as in easy to sync)
collection of definitions that are needed by separate Autotest
components, such as the RPC server and client.

This is a copy of the same commit applied on Autotest and every
new commit that touches "autotest.shared" should be replicated
here.

Signed-off-by: Cleber Rosa <crosa@redhat.com>